requirements.txt
fastapi==0.110.1
Jinja2==3.1.3
numpy==1.24.4
python-dotenv==1.0.1
supabase==2.4.1
uvicorn==0.29.0
python-multipart==0.0.9
pymupdf==1.24.4
openai==1.17.0
tiktoken==0.7.0